變更前處理圖片大小

+++ check_img.py
@@ -18,9 +18,10 @@ except OSError:
 
 def img_process(picture: str) -> Tuple[str, np.ndarray]:
     # 圖片前置處理
+    img_size = 300
     img = cv2.imread(picture, cv2.IMREAD_GRAYSCALE)
-    img = cv2.resize(img, (24, 24))
-    img = img.reshape(24, 24)
+    img = cv2.resize(img, (img_size, img_size))
+    img = img.reshape(img_size, img_size)
     return picture, img
 
 
@@ -83,35 +84,37 @@ name, del_count = 1, cdel_count  # 顯示在控制台
 # ex_img 是要比對的圖片(檢查過的)
 for un_img in uncheck_img:
     for ex_img in exist_img:
-        different = cv2.subtract(un_img.img_array, ex_img.img_array)
-        result = not np.any(different)
+        # 副檔名相同
+        if un_img.img_format == ex_img.img_format:
+            different = cv2.subtract(un_img.img_array, ex_img.img_array)
+            result = not np.any(different)
 
-        # 找到一樣的圖片
-        if result:
-            un_img.check = True
-            del_img = cv2.imread(un_img.img_name)
+            # 找到一樣的圖片
+            if result:
+                un_img.check = True
+                del_img = cv2.imread(un_img.img_name)
 
-            # 備份重複的圖片
-            shutil.copyfile(
-                un_img.img_name, f"./del_images/del_img_{del_count + 1}.{un_img.img_format}")
+                # 備份重複的圖片
+                shutil.copyfile(
+                    un_img.img_name, f"./del_images/del_img_{del_count + 1}.{un_img.img_format}")
 
-            # 刪除圖片
-            os.remove(un_img.img_name)
-            del_count += 1
-            break
+                # 刪除圖片
+                os.remove(un_img.img_name)
+                del_count += 1
+                break
 
     # 更改圖片名字
     if un_img.check == False:
         rename = f"LO_{cname}.{un_img.img_format}"
         os.rename(un_img.img_name, rename)
         exist_img.append(P(True, un_img.img_array, rename, un_img.img_format))
-        print(f"檢查第 {name} 張", end='\r', flush=True)
         cname += 1
-        name += 1
+    print(f"檢查第 {name} 張", end='\r', flush=True)
+    name += 1
 
 # 檔案尾數的紀錄
 write_file(cname, del_count)
 
-del_count = del_count - cdel_count
+del_count = del_count if del_count > 0 else 0
 print(
-    f"\n檢查完成!!\n新圖片的張數 : {name - 1}\n重複的張數 : {del_count if del_count > 0 else 0}")
+    f"\n檢查完成!!\n新圖片的張數 : {len(uncheck_img) - del_count}\n重複的張數 : {del_count}")
